this is refacoring done using the useContext
